fix(switch): restore correct on_toggle behavior and stabilize state handling
Replace problematic closures that captured and mutated switch state with callbacks that correctly follow the component API to avoid improper locking and toggle bugs. Update example showcases to use the new callback form. Also include related stability and cleanup changes: fix bottom sheet placement type, pass parent_constraint by value in column measurement, enforce fluid_glass instance limits and propagate buffer/bind_group creation failures with ?, remove redundant clippy allows, and simplify surface ripple matching.

+++ tessera-ui-basic-components/src/bottom_sheet.rs
@@ -241,7 +241,7 @@ fn place_bottom_sheet_if_present(
     let parent_height = input.parent_constraint.height.get_max().unwrap_or(Px(0));
     let current_is_open = state_for_measure.read().is_open;
     let y = compute_bottom_sheet_y(parent_height, child_size.height, progress, current_is_open);
-    input.place_child(bottom_sheet_id, PxPosition::new(Px(0), Px(y as i32)));
+    input.place_child(bottom_sheet_id, PxPosition::new(Px(0), Px(y)));
 }
 
 #[tessera]
